setup datagrid pagination, sorting and filtering

+++ server/modules/tasks/task.service.js
@@ -1,6 +1,6 @@
 const Task = require("./task.model");
 const { ResponseSuccess, ResponseError } = require("../../shared/response");
-const { ErrorsHandler } = require("../../utils");
+const { ErrorsHandler, DataGridHandler } = require("../../utils");
 const MessageBroker = require('../../config/MessageBroker')
 const { isManager } = require("../../utils/permissions-handler/PermissionsHandler");
 const { NOTIFICATION_QUEUE } = require("../../constants");
@@ -38,6 +38,12 @@ class TaskService {
                     message: `Invalid task data !`
                 })
 
+            if (data.summary.length > 2500)
+                return new ResponseError({
+                    status: 400,
+                    message: `Summary must not exceed 2500 characters !`
+                })
+
             const task = new Task({ ...data, createdBy });
 
             let result = await task.save();
@@ -60,42 +66,13 @@ class TaskService {
         }
     }
 
-    findAll = async (query, user) => {
-        try {
-            if (!isManager(user)) {
-                query.createdBy = user._id
-            }
-            const result = await Task.find(query)
-                .populate({
-                    path: 'createdBy',
-                    model: 'User',
-                    select: 'email username firstname lastname role'
-                });
-            if (!result) {
-                return new ResponseError({
-                    status: 404,
-                    message: `Task not found !`
-                })
-            }
-            return new ResponseSuccess({
-                status: 200,
-                content: result
-            })
-
-        } catch (err) {
-            return new ResponseError(ErrorsHandler.handle(err, `${SERVICE_NAME}:findAll`))
-        }
-    }
-
     findAllPaginated = async (
-        { page, limit, sortField, sortOrder, search },
+        { page, limit, filterModel, sortModel },
         user
     ) => {
         try {
-            let filter = {}
-            if (search && search.trim() !== "") {
-                filter = { title: { $regex: search, $options: 'i' } }
-            }
+            let filter = DataGridHandler.filterHandler(filterModel)
+            let sort = DataGridHandler.sortHadnler(sortModel)
             if (!isManager(user)) {
                 filter.createdBy = user._id
             }
@@ -106,7 +83,7 @@ class TaskService {
             let result = await Task.find(filter)
                 .skip((page - 1) * limit)
                 .limit(limit)
-                //.sort({ [sortField]: sortOrder })
+                .sort(sort)
                 .populate({
                     path: 'createdBy',
                     model: 'User',
@@ -136,34 +113,6 @@ class TaskService {
         }
     }
 
-    findById = async (id, user) => {
-        try {
-            let filter = { _id: id }
-            if (!isManager(user)) {
-                filter.createdBy = user._id
-            }
-            const result = await Task.findById(filter)
-                .populate({
-                    path: 'createdBy',
-                    model: 'User',
-                    select: 'email username firstname lastname role'
-                });
-            if (!result) {
-                return new ResponseError({
-                    status: 404,
-                    message: `Task not found !`
-                })
-            }
-            return new ResponseSuccess({
-                status: 200,
-                content: result
-            })
-
-        } catch (err) {
-            return new ResponseError(ErrorsHandler.handle(err, `${SERVICE_NAME}:findById`))
-        }
-    }
-
     update = async (id, data, user) => {
         try {
             const task = await Task.findById(id)
@@ -172,8 +121,6 @@ class TaskService {
                     status: 404,
                     message: "Task not found !"
                 })
-            console.log(task.createdBy)
-            console.log(user._id)
             if (task.createdBy != user._id)
                 return new ResponseError({
                     status: 403,
@@ -185,6 +132,12 @@ class TaskService {
                     message: "Task already performed !"
                 })
             const { title, summary, isPerformed, performedAt } = data
+
+            if (summary && summary.length > 2500)
+                return new ResponseError({
+                    status: 400,
+                    message: `Summary must not exceed 2500 characters !`
+                })
             let result = await Task.findOneAndUpdate({ _id: id },
                 { title, summary, isPerformed, performedAt },
                 { new: true });
@@ -211,6 +164,11 @@ class TaskService {
                 return new ResponseError({
                     status: 404,
                     message: "Task not found !"
+                })
+            if (!isManager(user))
+                return new ResponseError({
+                    status: 403,
+                    message: "Permission denied !"
                 })
             const result = await Task.deleteOne({ _id: id });
